requirements.txt
moderngl>=5.8.0
moderngl-window>=2.4.2
glfw>=2.6.4
numpy>=1.26.0
sounddevice>=0.4.6
mido>=1.3.0
python-rtmidi>=1.5.8
pyrr>=0.10.3
Pillow>=10.1.0
imgui-bundle>=1.0.0
PyYAML>=6.0
